# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexora AI Business OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /crm/leads:
    get:
      operationId: getLeads
      tags: [crm]
      summary: List leads
      parameters:
        - name: stage
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Leads list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [crm]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /crm/leads/{id}:
    get:
      operationId: getLead
      tags: [crm]
      summary: Get lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [crm]
      summary: Update lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [crm]
      summary: Delete lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /crm/contacts:
    get:
      operationId: getContacts
      tags: [crm]
      summary: List contacts
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [crm]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /crm/contacts/{id}:
    patch:
      operationId: updateContact
      tags: [crm]
      summary: Update contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [crm]
      summary: Delete contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Service Desk / Tickets
...
  /finance/transactions:
    get:
      operationId: getTransactions
      tags: [finance]
      summary: List transactions
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [finance]
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /field-service/orders:
    get:
      operationId: getWorkOrders
      tags: [field-service]
      summary: List work orders
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: technicianId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Work orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
    post:
      operationId: createWorkOrder
      tags: [field-service]
      summary: Create work order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
  /field-service/orders/{id}:
    patch:
      operationId: updateWorkOrder
      tags: [field-service]
      summary: Update work order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
...
  /automations:
    get:
      operationId: getAutomations
      tags: [automations]
      summary: List automations
      responses:
        "200":
          description: Automations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Automation"
    post:
      operationId: createAutomation
      tags: [automations]
      summary: Create automation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
  /automations/{id}:
    patch:
      operationId: updateAutomation
      tags: [automations]
      summary: Update automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
    delete:
      operationId: deleteAutomation
      tags: [automations]
      summary: Delete automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Lead:
      type: object
      required: [id, name, email, stage, value, probability, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        stage:
          type: string
        value:
          type: number
        probability:
          type: integer
        aiScore:
          type: ["integer", "null"]
        assignedTo:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    LeadInput:
      type: object
      required: [name, email, stage, value]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        stage:
          type: string
        value:
          type: number
        probability:
          type: integer
        assignedTo:
          type: ["string", "null"]
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        stage:
          type: string
        value:
          type: number
        probability:
          type: integer
        assignedTo:
          type: ["string", "null"]
...
    Contact:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
    ContactInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
    ContactUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        role:
          type: ["string", "null"]

    # Tickets
...
    Transaction:
      type: object
      required: [id, type, amount, description, category, date, status]
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        category:
          type: string
        date:
          type: string
        status:
          type: string
        reference:
          type: ["string", "null"]
    TransactionInput:
      type: object
      required: [type, amount, description, category, date]
      properties:
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        category:
          type: string
        date:
          type: string

    # Field Service
    WorkOrder:
      type: object
      required: [id, title, status, priority, technicianId, address, scheduledAt, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        technicianId:
          type: integer
        technicianName:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        address:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        scheduledAt:
          type: string
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    WorkOrderInput:
      type: object
      required: [title, priority, technicianId, address, scheduledAt]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
        technicianId:
          type: integer
        customerName:
          type: ["string", "null"]
        address:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        scheduledAt:
          type: string
    WorkOrderUpdate:
      type: object
      properties:
        status:
          type: string
        technicianId:
          type: integer
        scheduledAt:
          type: string
        completedAt:
          type: ["string", "null"]
...
    Automation:
      type: object
      required: [id, name, trigger, status, executionsCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        trigger:
          type: string
        status:
          type: string
        executionsCount:
          type: integer
        lastExecutedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    AutomationInput:
      type: object
      required: [name, trigger]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        trigger:
          type: string
    AutomationUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        trigger:
          type: string
        status:
          type: string