# lib/api-zod/src/generated/types/automationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexora AI Business OS API
 * OpenAPI spec version: 0.1.0
 */

export interface AutomationInput {
  name: string;
  /** @nullable */
  description?: string | null;
  trigger: string;
}