# lib/api-zod/src/generated/types/workOrderUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexora AI Business OS API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkOrderUpdate {
  status?: string;
  technicianId?: number;
  scheduledAt?: string;
  /** @nullable */
  completedAt?: string | null;
}